[v0.89][maintainability] Split obsmem_contract.rs by responsibility (Closes #1834)
